Happy numbers
A happy number is defined by the following process: Starting with any
positive integer, replace the number by the sum of the squares of its
digits, and repeat the process until the number equals 1 (where it will
stay), or it loops endlessly in a cycle which does not include 1.
Those numbers for which this process ends in 1 are happy numbers, while
those that do not end in 1 are unhappy numbers.
Find the first 8 happy numbers.